+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
+### v0.3.3
+
+**Update**
+- `getWallpapers` options replace to `WonderfulBingWallpaperOption`
+
 ### v0.3.1
 
 **Update**